Updates the property finder to find filters and borrows existing formatting
So to find the filters I actually open up the source code if I can find it.
There are a few more filters that have blocks that I am not finding. There
is always more to do in life.

I changed the formatting to look like more of the existing doc pages to make
it possibly easier to copy-n-paste out of it.

Signed-off-by: Franklin Webber <franklin@chef.io>